requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1

python-dotenv==1.0.1

ollama==0.4.4

pydantic==2.7.4
pydantic-settings==2.3.2

starlette==0.37.2

SQLAlchemy==2.0.31

watchdog==4.0.1

scipy==1.13.1
numpy==2.1.0

rich==13.7.1